requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
pillow==11.2.1
    # via -r requirements.in
protobuf==6.30.2
    # via -r requirements.in
pyotp==2.9.0
    # via -r requirements.in
pyzbar==0.1.9
    # via -r requirements.in
